namespace rework
everything available at top level

measurements get their own file
wrangle renamed to femap
